+++ assets/javascripts/kitten/hoc/media-queries.js
@@ -1,11 +1,16 @@
 import React, { Component } from 'react'
-import PropTypes from 'prop-types'
-import { createMatchMediaMax } from 'kitten/helpers/utils/media-queries'
+import {
+  createMatchMediaMax,
+  createMatchMedia,
+} from 'kitten/helpers/utils/media-queries'
 import {
   SCREEN_SIZE_XS,
   SCREEN_SIZE_S,
   SCREEN_SIZE_M,
 } from 'kitten/constants/screen-config'
+
+export const withMediaQueries = hocProps => WrapperComponent =>
+  mediaQueries(WrapperComponent, hocProps)
 
 export const mediaQueries = (WrappedComponent, hocProps = {}) =>
   class extends Component {
@@ -13,6 +18,16 @@ export const mediaQueries = (WrappedComponent, hocProps = {}) =>
       super(props)
 
       this.state = {}
+
+      const {
+        viewportIsMobile,
+        viewportIsSOrLess,
+        viewportIsTabletOrLess,
+        ...customProps
+      } = hocProps
+
+      this.custom = {}
+      this.customProps = customProps
 
       if (this.mobileMediaQueryEnabled()) {
         this.mqMobile = createMatchMediaMax(SCREEN_SIZE_XS)
@@ -28,6 +43,13 @@ export const mediaQueries = (WrappedComponent, hocProps = {}) =>
         this.mqTabletOrLess = createMatchMediaMax(SCREEN_SIZE_M)
         this.state = { ...this.state, viewportIsTabletOrLess: false }
       }
+
+      const customState = Object.keys(this.customProps).reduce(
+        (result, prop) => ({ ...result, [prop]: false }),
+        {},
+      )
+
+      this.state = { ...this.state, ...customState }
     }
 
     mobileMediaQueryEnabled = () =>
@@ -60,6 +82,13 @@ export const mediaQueries = (WrappedComponent, hocProps = {}) =>
         this.mqTabletOrLess.addListener(this.onTabletMQ)
         this.onTabletMQ(this.mqTabletOrLess)
       }
+
+      Object.keys(this.customProps).forEach(prop => {
+        this.setState({ [prop]: false })
+        this.custom[prop] = createMatchMedia(this.customProps[prop])
+        this.custom[prop].cb = event => this.setState({ [prop]: event.matches })
+        this.custom[prop].addListener(this.custom[prop].cb)
+      })
     }
 
     componentWillUnmount() {
@@ -74,6 +103,10 @@ export const mediaQueries = (WrappedComponent, hocProps = {}) =>
       if (this.mqTabletOrLess) {
         this.mqTabletOrLess.removeListener(this.onTabletMQ)
       }
+
+      this.customProps = Object.keys(this.customProps).forEach(prop =>
+        this.custom[prop].removeListener(this.custion[prop].cb),
+      )
     }
 
     onMobileMQ = event => {
@@ -89,13 +122,6 @@ export const mediaQueries = (WrappedComponent, hocProps = {}) =>
     }
 
     render() {
-      return (
-        <WrappedComponent
-          {...this.props}
-          viewportIsMobile={this.state.viewportIsMobile}
-          viewportIsSOrLess={this.state.viewportIsSOrLess}
-          viewportIsTabletOrLess={this.state.viewportIsTabletOrLess}
-        />
-      )
+      return <WrappedComponent {...this.props} {...this.state} />
     }
   }